+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/statictools/PiSDFFlattener.java
@@ -185,29 +185,6 @@ public class PiSDFFlattener extends PiMMSwitch<Boolean> {
   /**
    *
    */
-  public static void instantiateParameters(final AbstractActor actor, final AbstractActor copyActor,
-      final PiGraph resultPiGraph) {
-    // // Copy parameters
-    for (final Parameter p : actor.getInputParameters()) {
-
-      final List<ConfigInputPort> ports = actor.lookupConfigInputPortsConnectedWithParameter(p);
-      for (ConfigInputPort port : ports) {
-        final ConfigInputPort cip = (ConfigInputPort) copyActor.lookupPort(port.getName());
-        if (cip != null) {
-          final Parameter copy = PiMMUserFactory.instance.copyWithHistory(p);
-          final Dependency dep = PiMMUserFactory.instance.createDependency();
-          dep.setSetter(copy);
-          cip.setIncomingDependency(dep);
-          resultPiGraph.addDependency(dep);
-          resultPiGraph.addParameter(copy);
-        }
-      }
-    }
-  }
-
-  /**
-   *
-   */
   private void instantiateDependencies(final AbstractActor actor, final AbstractActor copyActor) {
     // // Copy parameters
 
@@ -228,19 +205,6 @@ public class PiSDFFlattener extends PiMMSwitch<Boolean> {
       }
     }
 
-    // for (final Parameter p : actor.getInputParameters()) {
-    //
-    // final EList<ConfigInputPort> ports = actor.lookupConfigInputPortsConnectedWithParameter(p);
-    // for (ConfigInputPort port : ports) {
-    // final ConfigInputPort cip = (ConfigInputPort) copyActor.lookupPort(port.getName());
-    // if (cip != null) {
-    // final Parameter parameter = this.param2param.get(p);
-    // final Dependency dep = PiMMUserFactory.instance.createDependency();
-    // dep.setSetter(parameter);
-    // cip.setIncomingDependency(dep);
-    // }
-    // }
-    // }
   }
 
   @Override